requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.0
python-multipart==0.0.6
sqlalchemy==2.0.34
alembic==1.13.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4